Updates from "gentleman hal":
There was a minor bug in the curve interpolation controller when cycle mode was used. Basically the order of the vectors being interpolated was slightly wrong resulting in jerky movement.

The submitted patch fixes it (it also adds a underscore to a class field in the quaternion controller to match ardor standards)

git-svn-id: http://ardorlabs.svn.cvsdude.com/Ardor3Dv1/trunk@1269 82abaa13-a93f-4eff-8987-b3ae9fc18050